Add accessor to the graph defined in that poperty. Since a property must be attached to a graph, there is no reason not to give access to that information.
ADD documentation of some fucntions,
There is a lot of functions undocumented here.
Furthermore it seems that all the public function used to 
manage push/pop should at least protected. Work should be done
on that interdace, since it is one of the central point to access to 
properies information.

David.