removed the GlHierarchyConvexHulls class. It is replaced by GlCompositeHierarchyManager (in tulip-qt)
initial import of the GlCompositeHierarchyManager class.*
It is intended to manage GlComposites of GlEntities following the graph/subgraph structure.
It can use any property as the name (defaults to the "name" property, and would only require to be passed a factory to the constructor to create any GlEntity.
As of now it only manages the ConvexHulls.

Bugfix from David to DrawingTools (caused the hulls to behave extremely strangely).

modified CMakeLists.txt to not build the olg GlHierarchyConvexHulls but the GlCompositeHierarchy instead.