add observed function in observable to prevent building of message when not necessary (idea could be apply in tlp3, to prevent non necessary function calls)
Succeed to simulate with one Event object the Tulip 3 behaviour of GraphObserver etc...
see the PointObserver in main.cpp for details.

Using that that system could be merged in Tulip 3 with small change of Observer code 
only need to add one line function :
    void update(const Event &event) {
        dynamic_cast<const [NameOfObservable]Event&>(event).call(*this);
    }

Observable will need more changes but there are easy and small. 

I Don't know if we will use that in Tlp4, but in term of pattern it is funny, something between the Observer and the Visitor. Tulip geeks, have a look and enjoy !!!