port of commit #6086 from branch 3_7
fixed a bug when setting data on a view:
the GlMainWidget deleted all layers, which in turn deleted cameras, which sent an event to the glQuadTreeLODCalculator, which rebuilt its camera and layer list.
But it rebuilt it using the list of layers from GlMainWidget which did not reflect the newly-deleted layers, thus accessing a deleted object.